# shared/api/openapi.yaml
openapi: 3.1.0
info:
  title: backend
  description: ''
  license:
    name: ''
  version: 0.1.0
paths:
  /auth/logout:
    post:
      tags: []
      operationId: logout
      responses:
        '204':
          description: Logged out
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
  /auth/oidc/callback:
    get:
      tags: []
      operationId: oidc_callback
      parameters:
      - name: code
        in: query
        required: true
        schema:
          type: string
      - name: state
        in: query
        required: true
        schema:
          type: string
      responses:
        '302':
          description: Redirect to frontend
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
  /auth/oidc/start:
    get:
      tags: []
      operationId: oidc_start
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                required: [redirect_url]
                properties:
                  redirect_url:
                    type: string
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
  /auth/refresh:
    post:
      tags: []
      operationId: refresh
      responses:
        '204':
          description: Refreshed
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
...
  /users/me:
    get:
      tags: []
      operationId: get_user_profile
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                required: [id, display_name, language_code, joined_at]
                properties:
                  display_name:
                    type: string
                  id:
                    type: string
                  joined_at:
                    type: string
                    format: date-time
                  language_code:
                    type: string
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - Bearer: []
    delete:
      tags: []
      operationId: remove_user
      responses:
        '204':
          description: Deleted
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - Bearer: []
    patch:
      tags: []
      operationId: update_user_profile
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name:
                  type:
                  - string
                  - 'null'
                  maxLength: 32
                language_code:
                  type:
                  - string
                  - 'null'
                  maxLength: 7
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                required: [id, display_name, language_code, joined_at]
                properties:
                  display_name:
                    type: string
                  id:
                    type: string
                  joined_at:
                    type: string
                    format: date-time
                  language_code:
                    type: string
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - Bearer: []
  /users/me/dashboard:
    get:
      tags: []
      operationId: get_dashboard
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                required: [ai_learning_summary, ai_learning_summary_updated_at, recent_seen_files]
                properties:
                  ai_learning_summary:
                    type: string
                  ai_learning_summary_updated_at:
                    type: string
                    format: date-time
                  recent_seen_files:
                    type: array
                    items:
                      type: object
                      required: [id, name, user_id, storage_type, storage_id, changed_at]
                      properties:
                        changed_at:
                          type: string
                          format: date-time
                        id:
                          type: string
                        name:
                          type: string
                        storage_id:
                          type: string
                        storage_type:
                          type: integer
                          format: int32
                          minimum: 0
                        user_id:
                          type: string
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - Bearer: []
...
  /users/me/sessions/{session_id}:
    delete:
      tags: []
      operationId: revoke_session
      parameters:
      - name: session_id
        in: path
        description: Session ID to revoke
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Revoked
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - Bearer: []
components:
  schemas:
    ErrorBody:
      type: object
      required: [code, description]
      properties:
        code:
          type: string
        description:
          type: string